FIX: Move mne.io.eyelink.read_eyelink_calibraiton to mne.preprocessing.eyetracking
Moved the function to the mne.preprocessing.eyetracking.calibration py file

Updated the documentation of the function in doc
