+++ mne/io/eyelink/__init__.py
@@ -5,4 +5,4 @@
 #
 # License: BSD-3-Clause
 
-from .eyelink import read_raw_eyelink, read_eyelink_calibration
+from .eyelink import read_raw_eyelink
