+++ mne/io/eyelink/eyelink.py
@@ -10,7 +10,6 @@ from datetime import datetime, timezone, timedelta
 from pathlib import Path
 
 import numpy as np
-from ._utils import _parse_calibration
 from ..constants import FIFF
 from ..base import BaseRaw
 from ..meas_info import create_info
@@ -288,43 +287,6 @@ def _find_overlaps(df, max_time=0.05):
 
 
 @fill_doc
-def read_eyelink_calibration(
-    fname, screen_size=None, screen_distance=None, screen_resolution=None
-):
-    """Return info on calibrations collected in an eyelink file.
-
-    Parameters
-    ----------
-    fname : path-like
-        Path to the eyelink file (.asc).
-    screen_size : array-like of shape (2,)
-        The width and height (in meters) of the screen that the eyetracking
-        data was collected with. For example ``(.531, .298)`` for a monitor with
-        a display area of 531 x 298 mm. Defaults to None.
-    screen_distance : float
-        The distance (in meters) from the participant's eyes to the screen.
-        Defaults to None.
-    screen_resolution : array-like of shape (2,)
-        The resolution (in pixels) of the screen that the eyetracking data
-        was collected with. For example, ``(1920, 1080)`` for a 1920x1080
-        resolution display. Defaults to None.
-
-    Returns
-    -------
-    calibrations : list
-        A list of :class:`mne.preprocessing.eyetracking.Calibration` instances, one for
-        each eye of every calibration that was performed during the recording session.
-    """
-    fname = _check_fname(fname, overwrite="read", must_exist=True, name="fname")
-
-    if not fname.exists():
-        raise FileNotFoundError(f"Could not find file {fname}")
-    logger.info("Reading calibration data from {}".format(fname))
-    lines = fname.read_text(encoding="ASCII").splitlines()
-    return _parse_calibration(lines, screen_size, screen_distance, screen_resolution)
-
-
-@fill_doc
 def read_raw_eyelink(
     fname,
     preload=False,
@@ -421,6 +383,8 @@ def read_raw_eyelink(
         gap_desc=gap_description,
     )
     if return_calibration:
+        from ...preprocessing.eyetracking import read_eyelink_calibration
+
         calibrations = read_eyelink_calibration(
             fname=fname,
             screen_size=screen_size,
