+++ mne/preprocessing/eyetracking/__init__.py
@@ -5,4 +5,4 @@
 # License: BSD-3-Clause
 
 from .eyetracking import set_channel_types_eyetrack
-from .calibration import Calibration
+from .calibration import Calibration, read_eyelink_calibration
